Old Watchtower - Minor Forest Addition (#3424)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds the Forest Watchtower, previously used for ensuring the road
towards Vanderlin were safe. Now, it has been abandoned with it's
previous occupants still stumbling around inside in their eternal duty.
Consists of three levels, with varying levels of danger on each.
Skeletons can be changed if balance is an issue.

<img width="832" height="704" alt="image"
src="https://github.com/user-attachments/assets/0dc75559-22e6-44eb-ac21-11f42a5cd89f"
/>

## Why It's Good For The Game

Adds more variety to the region, it's an otherwise desolate part of the
forest so I would think that it would be a decent place for it.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
map: Adds the Old Watchtower to the Forest.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->