Great Axe & Double Head Great Axe Balance + Made them craftable (#3355)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Greataxe got balanced towards being a heavy hitting weapon but being
decent/bad to parry with, also being easy to dodge, i tried to make the
weapons different enough so it is not a polearm with a different sprite


Greataxe boons:

- 2 Reach on Chop and Cut intent

- Chop has the Same AP from the Halberds Chop intent

- Good damage for both Chop and Cut intents

Greataxe flaws:

- Big drain if you miss for both Cut and Chop intents

- Easy to dodge of

- Average to Parry with

- Needs to charge for the Chop and Cut intent

Double head Greataxe boons:

- Same greataxe boons

- More Damage on both Chop and Cut intents than the normal greataxe

- Higher blade integrity (Takes more time to lose sharpness so helps
with delimb)

Double head Greataxe flaws:

- Increased drain if you miss both Cut and Chop intents

- Higher charge for the Chop and Cut Intents

- Very easy to dodge of


## Why It's Good For The Game

New craftable weapon.

Balance towards the classes that already get it is, it is a 2 reach
weapon that trades damage for defense, while Polearms got less damage it
is better at defense and Greatswords is kinda in the midle. (Exceptions
being the uncraftable weapons)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Great Axes and Double Head Great Axes are now craftable
bal: Greataxes and Double Head Great Axes got a limit of 30 force
(exception being the graggar axe dreamer variant)
bal: Greataxes and Double Head Great Axes ,got reach 2 in chop and cut
doing great damage but got are worse to parry with and easier to dodge
of with a bigger stamina drain on miss. (Check the PR for more detailed
info)
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->