Removes vagrant/l, makes vagrant randomized (#3447)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The title makes no sense
So basically
There is a vagrant shirt, and a vagrant/l shirt, same for tights.
Vagrant shirts/tights have one torn sleeve, the left variant literally
has a left torn sleeve instead of the right one.
That's it.
And INSTEAD OF MAKING IT RANDOM ON INITIALIZE
someone decided to just randomly pick between the two instead of you
know, making the shirts/tights randomize it.

Also cleaned the random skeleton warrior code a bit by moving all the
same items (coif, pants, etc.) before the switch.
Somewhat unrelated for the PR but they used the vagrant tights/pants and
I basically noticed it.

No changelog because nothing happened that the players would care about.

## Why It's Good For The Game

Makes things more consistent, basically

## Changelog

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->