Persistence server jobs (#3462)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Big thanks to wgtjunior, grunguss and cheffie for helping me with this.
Appreciate it.

I was tasked to make jobs (and other stuff) for the persistence server,
here they are.
Todos:
- [x] Make the jobs.
- [x] Make some sort of toggle/option so only these jobs are selectable.
- [x] modular experience, someway to change experience gain rate.

## Why It's Good For The Game

So the persistence server has jobs and stuff.

## Changelog

No changelog because I don't think it's needed

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->