Tweaks: Influences, ascendants, personal objectives and roleplay triumph buys (#3517)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Minor influences balancing
- Smaller god ranking influence bonus/penalty
- Personal objectives influence bonus normalized to 20 for all
objectives
- Scheduled events now run in 2 minutes instead of 3 (to avoid
situations where storyteller managed to change 3 times in the meantime)
- Slightly easier take pain and listen to dead's whispers objectives
- Discounted my previously added triumph buys a little, as these are
basically only to enhance roleplay, which players probably don't want to
spent too much on compared to more useful ones

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->