A few vampires fixes and changes (#3518)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Fixes Demonic Curses not working on someone without a family

- FIxes Blood runes giving the user blood instead of consuming it

- Fixes Punish Minion not working and expands it to include all members
of the clan including mortals.

- Disables the "move" button from the runes UI.

- Demonic Fireball now uses blood instead of mana and increases the
cooldown

- Gives obfuscation users silent steps while they move


## Why It's Good For The Game

- Fix

- I had to disable it since the UI just tps away once you click and ppl
were accidentally doing that all the time and losing the ability to use
blood runes forever, it is better this way until me or someone fixes it

- It used mana for some reason and it is given by the demonic coven
being basically an one time use since 99% of the vampires including the
VL are most likely not mages, atleast this is now useful.

- Just like the WoD obfuscation, it is silent, the coven power is
supposed to make so you can't notice that person anymore and it is kinda
hard to do that when you can hear their steps very loudly, this will
also make it more useful. (Any combat action disables it)


## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Demonic Curses now can be applied to members without a family.
fix: Blood runes now uses vitae instead of giving it.
fix: Vampire Lord Punish minion verb now works.
del: Comments the "move" button from the blood runes UI until it is
fixed.
bal: The demonic fireball given by the demonic coven now uses blood
instead of mana and has a cooldown of 60 seconds.
bal: Moving while using obfuscation now makes the user steps silent.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->