Expands Rejected Royal's allowed species (#3542)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds more races to the Rejected Royal Wretch that I felt would fit.
Specifically half drow, full drow, half orcs, and Tieflings

Half-drow because obvious reasons, heritage disputes.

Full drow for a similar reason, maybe a little drow in the family
skipped a generation or two.

Half-orcs because someone can become a half-orc in their lifespan or be
born as one because someone in their family was one once.

Tieflings because people have Tiefling children spontaneously with no
apparent reasoning beyond Baotha deciding to troll you.

'How'd they not all just get killed at birth?' Well, that's for the
roleplayers to decide. Maybe the King didn't want to bash in their child
Tiefling's head and opted to just kick them out.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

The previous setup for rejected royal was only royal races (same as King
pretty much), which was boring. Learning that the humen king has a
half-drow child by the King is interesting, learning that he has a humen
child isn't. In short, more dramatic stuff happens because of the race
lock being lessened.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added half-drow, drow, half-orcs, and tieflings to Rejected Royal
Wretch
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->