Remove Age modifier for Maniac/Lich and Trey Liam change (#3298)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Buff Maniac by removing the age_innate modifier on spawn and make trey
into a actual human.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Maniac Youngling is rather weak and they're forced into only doing
things that are stealth-oriented like killing peoples in their sleep to
harvest organs due to a lack of Con and Str,
- They can't take blows
- They are not able to knock peoples down by using their fists when
regular maniac can.
- they are also unable to kick open doors which mean they are forced to
either destroy it by hands or find another way.
We literally shouldn't be removing access to gimmicks just because you
rolled maniac as a youngling, it's legit common sense, Despite what
peoples say, Maniac is literally a fragger antag, they're a one-man only
antag that are literally forced to murder more than 4 peoples, so we
shouldn't handicap someone and making it harder for them just because
they rolled a antag.

And I noticed that trey liam was a humen instead of a human, so we fix
that, yk yk
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
balance: Youngling/Old Maniac/lich have now the same stats as their
adult counterpart.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->