The Peddlering pt. 2 (#3452)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds two new overlays for unique peddlers around the map. A green tint
for the FARMHAND and a red tint for the newly named BUTCHER peddler.
Also applies the existing merch overlay to the newly named SHOPHAND
peddler for the merchant. (This was only used for steward previously,
for some reason.)
Fixes an issue with my previous peddler PR where the peddlers were not
properly calling out renamed items. Apothecary and Soilson rejoice.
Also edited the keys held by both the steward and innkeeper's peddlers.
Steward held three shop keys for shops that no longer exist, now only
holds shop 4 at an increased price of 20 mammons. It also holds a key
for house 7, which does not exist on Vanderlin, leaving that until we
make distinct peddlers for maps. Innkeep held three keys for the
thatchwood inn, making them utterly useless and a money trap. Corrected
the type and removed all of the keys save for medium room 1. It made no
sense to have so many copies of keys for private rooms available.
![ss+(2025-09-22+at+11 50
14)](https://github.com/user-attachments/assets/449db42e-c233-4cfd-9b00-f278bb796705)
New soilson and butcher overlays with the existing merchant, apothecary,
and inn.

![ss+(2025-09-22+at+11 54
09)](https://github.com/user-attachments/assets/c6411467-7b8b-483d-a609-eb4fc3b4e4e5)
Peddlers properly hawking a named item.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More flavor and better visibility for peddlers.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

fix: #3188
fix: #3348 

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: More filled_overlay sprites for peddlers
fix: Innkeep and Steward peddler now hold keys that can actually be used
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>